refactor: extract volume filter
