+++ volatilitySelector.js
@@ -3,11 +3,7 @@ const { TOP_N_PAIRS, DEBUG_LOG_LEVEL, VOLUME_FILTER } = require('./config');
 const { pruneObsoleteSymbols } = require('./utils/pruneCache');
 const { loadFuturesSymbols, isFuturesTradable, hasFuturesData } = require('./futuresSymbols');
 const { verboseLog, basicLog } = require('./utils/logger');
-
-function calcRecentVolumeUSD(candles = [], count = 5) {
-  if (!Array.isArray(candles) || candles.length === 0) return 0;
-  return candles.slice(-count).reduce((sum, c) => sum + c.volume * c.close, 0);
-}
+const { filterSymbolsByVolume } = require('./utils/volumeFilter');
 
 async function getTopVolatilePairs(candleCache) {
   try {
@@ -40,23 +36,7 @@ async function getTopVolatilePairs(candleCache) {
     topSymbols = topSymbols.filter(p => isFuturesTradable(p.symbol));
 
     if (VOLUME_FILTER?.ENABLED) {
-      const removed = [];
-      topSymbols = topSymbols.filter(p => {
-        const candles = candleCache?.[p.symbol]?.['5m'] || [];
-        const volumeUSD = calcRecentVolumeUSD(candles, 5);
-        if (volumeUSD < VOLUME_FILTER.MIN_VOLUME_5M_USD) {
-          removed.push({ symbol: p.symbol, volumeUSD });
-          return false;
-        }
-        return true;
-      });
-
-      removed.forEach(r => {
-        if (DEBUG_LOG_LEVEL !== 'none') {
-          const vol = r.volumeUSD.toLocaleString(undefined, { maximumFractionDigits: 0 });
-          basicLog(`[INFO] 🔇 Пропуск ${r.symbol} — объём $${vol} за 5 минут ниже порога $${VOLUME_FILTER.MIN_VOLUME_5M_USD}`);
-        }
-      });
+      topSymbols = filterSymbolsByVolume(topSymbols, candleCache);
     }
 
     if (DEBUG_LOG_LEVEL === 'verbose' && excluded.length) {
